back_end/requirements.txt
alembic==1.9.4
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.1
cachelib==0.10.2
click==8.1.3
flasgger==0.9.5
Flask==2.2.3
Flask-Cors==3.0.10
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-Session==0.4.0
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
importlib-metadata==6.0.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
Mako==1.2.4
MarkupSafe==2.1.2
mistune==2.0.5
mysqlclient==2.1.1
packaging==23.0
pkgutil_resolve_name==1.3.10
pycodestyle==2.10.0
pyrsistent==0.19.3
python-dotenv==1.0.0
PyYAML==6.0
redis==4.5.1
six==1.16.0
SQLAlchemy==2.0.4
tomli==2.0.1
typing_extensions==4.5.0
Werkzeug==2.2.3
zipp==3.14.0
